+++ changelog.md
@@ -21,6 +21,7 @@
 * Socializer is now more tolerant to internet issues. When attempting to create a wall post, topic or send a chat message, if the data is unable to be posted to VK, socializer will allow you to try to post it again, giving you the opportunity to edit or copy the text of the post in case you want to save it for later. Later change will also be introduced when posting comments, too.
 * Less confidential user data will be send to the logs, so it will be much safer to pass logs publicly.
 * automatic update checks will be disabled if using socializer from the source code.
+* it is possible to post in an user's wall by using the post button located next to the user, in people buffers. This applies only to online users and list of friends.
 
 ## changes in Versions 0.21 and 0.22 (14.07.2019)